Updating Zipcode, BlockGroup models

+++ main/models.py
@@ -2,16 +2,16 @@ from django.contrib.gis.db import models
 
 class Zipcode(models.Model):
     zipcode = models.CharField(max_length=5)
-    area_land = models.IntegerField() # in m^2
-    area_water = models.IntegerField() # in m^2
+    area_land = models.BigIntegerField() # in m^2
+    area_water = models.BigIntegerField() # in m^2
     mpoly = models.MultiPolygonField()
     def __str__(self):
         return self.zipcode
 
 class BlockGroup(models.Model):
     geoid = models.CharField(max_length=12)
-    area_land = models.IntegerField() # in m^2
-    area_water = models.IntegerField() # in m^2
+    area_land = models.BigIntegerField() # in m^2
+    area_water = models.BigIntegerField() # in m^2
     mpoly = models.MultiPolygonField()
     def __str__(self):
         return self.geoid
